Less warnings
git-svn-id: file:///Users/tack/GecodeGitMigration/gecode-svn-mirror/gecode/trunk@8279 e85b7adc-8362-4630-8c63-7469d557c915

+++ gecode/int/support-values.hpp
@@ -188,7 +188,7 @@ namespace Gecode { namespace Int {
   template<class View, class A>
   forceinline int
   SupportValues<View,A>::Unsupported::val(void) const {
-    return rp->min+(p-rp->pos);
+    return static_cast<int>(rp->min+(p-rp->pos));
   }
 
   template<class View, class A>
